Add file from upload
This is a Jupyter notebook that I exported from Hex (which is the IDE I use). In this file, I pull historical bitcoin price data from Kraken's REST API then use it to plot the price of bitcoin on Kraken. The formatting is a bit wonky because it's exported from Hex, but you probably get the general idea. This is the first time I've pulled data from an API! Exciting.